Start database development
Took about an hour or so to add the database into c#, but it has taken me 3 hours to figure out how to get the sql data to work. After around 2.5 hours I managed to get it into a viewable database, but getting a connection with it is almost impossible. I will at a later point manually convert the database into a sqlite db to lessen the time spent on the sql file importing